+++ js/data/model.js
@@ -1,8 +1,23 @@
-const maxLives = 3;
-const maxQuestionNum = 9;
+/**
+ * game data structure
+ * @type {object}
+ */
+export const game = {
+  currentQuestion: 0,
+  lives: 3,
+  timer: 0,
+  maxTime: 120,
+  maxQuestionNum: 9
+};
 
+/**
+ * sets lives and returns new game object
+ * @param {object} gameObj
+ * @param {number} lives
+ * @return {object}
+ */
 export const setLives = (gameObj, lives) => {
-  if (lives < 0 || lives > maxLives) {
+  if (lives < 0 || lives > game.lives) {
     throw new RangeError(`Number of lives can't be negative or > 3`);
   }
 
@@ -11,6 +26,13 @@ export const setLives = (gameObj, lives) => {
   });
 };
 
+/**
+ * sets time for current game
+ * and returns new game object
+ * @param {object} gameObj
+ * @param {number} time
+ * @return {object}
+ */
 export const setTime = (gameObj, time) => {
   if (time < 0) {
     throw new RangeError(`Time can't be negative`);
@@ -21,8 +43,15 @@ export const setTime = (gameObj, time) => {
   });
 };
 
+/**
+ * sets current game question
+ * and returns new game object
+ * @param {object} gameObj
+ * @param {number} questionNum
+ * @return {object}
+ */
 export const setCurrentQuestion = (gameObj, questionNum) => {
-  if (questionNum < 0 || questionNum > maxQuestionNum) {
+  if (questionNum < 0 || questionNum > game.maxQuestionNum) {
     throw new RangeError(`Questions can't be negative or > 9`);
   }
