+++ js/template-welcome.js
@@ -1,5 +1,4 @@
 import * as dom from 'dom-helpers';
-import {game} from './data/game-data';
 import {gameStart} from 'game';
 
 export default (inputData) => {
@@ -39,7 +38,7 @@ export default (inputData) => {
    * when clicked, the first question is rendered
    */
   playButton.addEventListener('click', () => {
-    gameStart(game);
+    gameStart();
   });
 
   return element;
